Move to separate messages folder and add message_v2
