+++ bindings_ffi/src/message.rs
@@ -10,7 +10,7 @@ use xmtp_content_types::{
     wallet_send_calls::{WalletCall, WalletCallMetadata, WalletSendCalls},
 };
 use xmtp_db::group_message::{DeliveryStatus, GroupMessageKind};
-use xmtp_mls::groups::decoded_message::{
+use xmtp_mls::messages::decoded_message::{
     DecodedMessage, DecodedMessageMetadata, MessageBody, Reply as ProcessedReply, Text,
 };
 use xmtp_proto::xmtp::mls::message_contents::content_types::{
