+++ xmtp_mls/src/groups/message_list.rs
@@ -1,20 +1,10 @@
-use std::collections::HashMap;
-
-use hex::ToHexExt;
-use xmtp_db::DbQuery;
-use xmtp_db::group_message::{
-    ContentType as DbContentType, MsgQueryArgs, RelationCounts, RelationQuery, StoredGroupMessage,
-};
+use xmtp_db::group_message::{ContentType as DbContentType, MsgQueryArgs};
 
 use crate::context::XmtpSharedContext;
-use crate::groups::decoded_message::{DecodedMessage, MessageBody};
 use crate::groups::{GroupError, MlsGroup};
+use crate::messages::decoded_message::DecodedMessage;
+use crate::messages::enrichment::enrich_messages;
 use xmtp_db::prelude::QueryGroupMessage;
-
-// Mapping of reactions, keyed by the ID of the message being reacted to.
-type ReactionMap = HashMap<Vec<u8>, Vec<DecodedMessage>>;
-// Mapping of referenced messages, keyed by ID
-type ReferencedMessageMap = HashMap<Vec<u8>, DecodedMessage>;
 
 impl<Context> MlsGroup<Context>
 where
@@ -29,95 +19,8 @@ where
         let initial_messages =
             conn.get_group_messages(&self.group_id, &filter_out_reactions_from_query(query))?;
 
-        let initial_message_ids: Vec<&[u8]> =
-            initial_messages.iter().map(|m| m.id.as_ref()).collect();
-
-        let reference_ids: Vec<&[u8]> = initial_messages
-            .iter()
-            .filter_map(|m| m.reference_id.as_deref())
-            .collect();
-
-        let mut relations =
-            get_relations(conn, &self.group_id, &initial_message_ids, &reference_ids)?;
-
-        let messages: Vec<DecodedMessage> = initial_messages
-            .into_iter()
-            .filter_map(|stored_message| {
-                let mut decoded = DecodedMessage::try_from(stored_message)
-                    .inspect_err(|err| tracing::warn!("Failed to decode message {:?}", err))
-                    .ok()?;
-
-                decoded.reactions = relations
-                    .reactions
-                    .remove(&decoded.metadata.id)
-                    .unwrap_or_default();
-
-                decoded.num_replies = relations
-                    .reply_counts
-                    .get(&decoded.metadata.id)
-                    .cloned()
-                    .unwrap_or(0);
-
-                if let MessageBody::Reply(mut reply_body) = decoded.content {
-                    let _ = hex::decode(&reply_body.reference_id)
-                        .inspect_err(|err| {
-                            tracing::warn!("could not parse reference ID as hex: {:?}", err)
-                        })
-                        .inspect(|id| {
-                            let in_reply_to = relations.referenced_messages.get(id).cloned();
-                            reply_body.in_reply_to = in_reply_to.map(Box::new);
-                        });
-                    decoded.content = MessageBody::Reply(reply_body);
-                }
-
-                Some(decoded)
-            })
-            .collect();
-
-        Ok(messages)
+        Ok(enrich_messages(conn, &self.group_id, initial_messages)?)
     }
-}
-
-fn get_relations(
-    conn: impl DbQuery,
-    group_id: &[u8],
-    message_ids: &[&[u8]],
-    reference_ids: &[&[u8]],
-) -> Result<GetRelationsResults, GroupError> {
-    if message_ids.is_empty() {
-        return Ok(GetRelationsResults {
-            reactions: HashMap::new(),
-            referenced_messages: HashMap::new(),
-            reply_counts: HashMap::new(),
-        });
-    }
-
-    let reactions_relations_query = RelationQuery::builder()
-        .content_types(Some(vec![DbContentType::Reaction]))
-        .build()
-        .unwrap_or_default();
-
-    let replies_count_query = RelationQuery::builder()
-        .content_types(Some(vec![DbContentType::Reply]))
-        .build()
-        .unwrap_or_default();
-
-    let reactions = conn.get_inbound_relations(group_id, message_ids, reactions_relations_query)?;
-    let referenced_messages = conn.get_outbound_relations(group_id, reference_ids)?;
-    let reply_counts =
-        conn.get_inbound_relation_counts(group_id, message_ids, replies_count_query)?;
-
-    Ok(GetRelationsResults {
-        reactions: get_reactions(reactions),
-        referenced_messages: get_referenced_messages(referenced_messages),
-        reply_counts,
-    })
-}
-
-struct GetRelationsResults {
-    reactions: ReactionMap,
-    referenced_messages: ReferencedMessageMap,
-    reply_counts: RelationCounts,
 }
 
 fn filter_out_reactions_from_query(query: &MsgQueryArgs) -> MsgQueryArgs {
@@ -135,52 +38,12 @@ fn filter_out_reactions_from_query(query: &MsgQueryArgs) -> MsgQueryArgs {
     new_query
 }
 
-fn get_referenced_messages(messages: HashMap<Vec<u8>, StoredGroupMessage>) -> ReferencedMessageMap {
-    messages
-        .into_iter()
-        .filter_map(|(id, stored_message)| {
-            let message_id = id.clone();
-            DecodedMessage::try_from(stored_message)
-                .inspect_err(|err| {
-                    tracing::warn!(
-                        "Failed to decode reply root message with ID {} {:?}",
-                        message_id.encode_hex(),
-                        err
-                    );
-                })
-                .map(|decoded| (id, decoded))
-                .ok()
-        })
-        .collect()
-}
-
-fn get_reactions(messages: HashMap<Vec<u8>, Vec<StoredGroupMessage>>) -> ReactionMap {
-    messages
-        .into_iter()
-        .map(|(id, reaction_messages)| {
-            let mapped_reactions: Vec<DecodedMessage> = reaction_messages
-                .into_iter()
-                .filter_map(|stored_msg| {
-                    DecodedMessage::try_from(stored_msg)
-                        .inspect_err(|err| {
-                            tracing::warn!(
-                                "Failed to decode message categorized as Reaction: {:?}",
-                                err
-                            );
-                        })
-                        .ok()
-                })
-                .collect();
-            (id, mapped_reactions)
-        })
-        .collect()
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
     use crate::builder::ClientBuilder;
     use crate::groups::QueryableContentFields;
+    use crate::messages::decoded_message::MessageBody;
     use hex::ToHexExt;
     use xmtp_common::time::now_ns;
     use xmtp_content_types::ContentCodec;
@@ -188,7 +51,9 @@ mod tests {
     use xmtp_content_types::text::TextCodec;
     use xmtp_cryptography::utils::generate_local_wallet;
     use xmtp_db::Store;
-    use xmtp_db::group_message::{ContentType as DbContentType, DeliveryStatus, GroupMessageKind};
+    use xmtp_db::group_message::{
+        ContentType as DbContentType, DeliveryStatus, GroupMessageKind, StoredGroupMessage,
+    };
     use xmtp_proto::xmtp::mls::message_contents::content_types::ReactionAction;
     use xmtp_proto::xmtp::mls::message_contents::{ContentTypeId, EncodedContent};
