+++ xmtp_mls/src/groups/mod.rs
@@ -1,6 +1,5 @@
 pub mod commit_log;
 pub mod commit_log_key;
-pub mod decoded_message;
 pub mod device_sync;
 pub mod device_sync_legacy;
 pub mod disappearing_messages;
