+++ xmtp_mls/src/lib.rs
@@ -8,6 +8,7 @@ pub mod groups;
 pub mod identity;
 pub mod identity_updates;
 mod intents;
+pub mod messages;
 pub mod mls_store;
 mod mutex_registry;
 pub mod subscriptions;
