+++ xmtp_mls/src/messages/mod.rs
@@ -1,0 +1,2 @@
+pub mod decoded_message;
+pub mod enrichment;
